default dataset to 'files' for file write stream and listDatasets
I want to fix a small bug where `dat datasets` shows the `files`
dataset while `dat status` doesn’t count the `files` dataset. Since
people are starting to build on dat using js, I figure it’d be good to
normalize this api a bit.

With this change, when user1 builds a dat with the js api using
fileWriteStream, it’ll default to the `files` dataset — then someone
else can `dat clone` and be able to `dat read` as if it was added with
the cli.